requirements.txt
openai>=1.0.0
fastapi
uvicorn
python-dotenv
chromadb
twilio
pydantic
requests
python-multipart
